fixed broken tests due to variable JIT gucs in Explain Settings field
Modified gucs can be showed in the Settings field of Explain output. The field
is ignored in common EXPLAIN_FORMAT_TEXT by gpdiff tools. But for unaligned
mode output and other format like YAML, The Settings field is token into
account. As the value of JIT gucs varies among different tests and are not
the test point, this patch filtered out them in Settings field of Explain
output.